19/5/22

+++ 6a.sh
@@ -12,21 +12,26 @@ then
 		mkdir ~/mydir
 	fi	
 	for k in $*
-	do
-			find $HOME -name "$k">files
-			while read line
-			do
+	do	
+			if [ -f $k ]
+			then
+				find $HOME -name "$k">files
+				while read line
+				do
 			
-				num=`echo $line | grep -o "/" | wc -l`
-				num=`expr $num + 1`
-				file=`echo $line | cut -d "/" -f $num`
-				cat $line
-				echo "--------------------------------------------------------------"
-				if [ `ls ~/mydir | grep $file | wc -l` -eq 0 ]
-				then
-					cp $line ~/mydir 
-				fi
-			done<files
+					num=`echo $line | grep -o "/" | wc -l`
+					num=`expr $num + 1`
+					file=`echo $line | cut -d "/" -f $num`
+					cat $line
+					echo "--------------------------------------------------------------"
+					if [ `ls ~/mydir | grep $file | wc -l` -eq 0 ]
+					then
+						cp $line ~/mydir 
+					fi
+				done<files
+			else
+				echo "File doesn't exists"
+			fi
 	done
 else
 	echo "Give arguments"
